router: install and configure the basics of react-router
React-router was installed and configured in the application. There were created
four initial basic routes, /, /create, /edit and /help.

HisotoryAPIFallback was introduced to tell the server to send back the index.html file
whenever there was a 404 message (page not found).

Also 'exact' was used to differentiate the '/' from all the children urls.

Signed-off-by: Daniel Kreling <dbkreling@gmail.com>